+++ ChangeLog.md
@@ -14,6 +14,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ### Added
 
+- Added debugging of slicing/subsetting arrays
+
 ### Removed
 
 ### Fixed